requirements.txt
tiktoken
python-docx